webapp/requirements.txt
# pandas==1.5.3
# numpy==1.24.2
django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.4.0
django-cors-headers==3.14.0
# transformers==4.27.4
# datasets==2.11.0
# torch==2.0.0
# faiss-cpu==1.7.3
# git-lfs==1.6
django-environ==0.9.0
requests
# pandas
# matplotlib
# transformers
# openai
# git+https://github.com/rwightman/pytorch-image-models.git
# Pillow
# apex
# functorch